backend/requirements.txt
Django==5.0.3
celery==5.3.6
redis==5.0.3
django-auth-ldap==4.6.0
djangorestframework==3.15.1
pymorphy2==0.9.1
python-ldap==3.4.4
ldap3==2.9.1
django-python3-ldap==0.15.8
django-cors-headers==4.4.0
openpyxl==3.1.2
psycopg2==2.9.9
requests==2.32.3
django-cleanup==9.0.0
docxtpl==0.19.1
holidays==0.79